restructure with node-liftoff for versioned cli
