+++ package.json
@@ -9,6 +9,7 @@
     "graceful-fs": "~2.0.1",
     "jspm-github": "~0.6.0",
     "jspm-npm": "~0.6.0",
+    "liftoff": "^0.13.2",
     "minimatch": "^0.2.14",
     "mkdirp": "~0.3.5",
     "ncp": "~0.6.0",
